+++ docs/CHANGELOG.rst
@@ -22,6 +22,8 @@ Added
 - Support Python 3.7
 - Support forward and reverse many-to-one relations in Elasticsearch
 - Add ``collection_names`` field to ``DataSerializer``
+- Add test methods  to ``ProcessTestCase`` that assert directory structure and
+  content: ``assertDirExists``, ``assertDir``, and ``assertDirStructure``
 
 
 ===================